astro-observe-mcp/pyproject.toml
[project]
name = "astro-observe-mcp"
dynamic = ["version"]
description = "A FastMCP server for Astro Cloud Observability API - search, list, and retrieve catalog assets"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.1.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
astro-observe-mcp = "astro_observe_mcp.__main__:main"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"
fallback-version = "0.0.0+dev"
raw-options = { root = "..", tag_regex = "^astro-observe-mcp-(?P<version>.*)$", git_describe_command = ["git", "describe", "--tags", "--match", "astro-observe-mcp-*"] }

[tool.hatchling.build.targets.wheel]
packages = ["src/astro_observe_mcp"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.12.0",
    "pytest-httpx>=0.30.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
extend-select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "ISC",    # implicit string concatenation
    "LOG",    # flake8-logging
    "TC",     # type-checking imports
    "G",      # flake8-logging-format
    "RET",    # flake8-return
    "PIE",    # flake8-pie
    "PT",     # flake8-pytest-style
    "RUF",    # Ruff-specific rules
    "ASYNC",  # flake8-async
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "ARG002", # unused method argument
    "RUF012", # mutable class attributes
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = ["ARG001", "ARG002", "PT011"]
